% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrendfilters.R, R/plotmethods.R,
%   R/printmethods.R
\name{hfilter}
\alias{hfilter}
\alias{plot.hfilter}
\alias{print.hfilter}
\title{Filter proposed by Hamilton (2018, doi:10.1162/rest_a_00706) to separate trend
and cycle components of a series.}
\usage{
hfilter(y, h = 24, p = 12)

\method{plot}{hfilter}(x, ...)

\method{print}{hfilter}(x, ..., digits = 3)
}
\arguments{
\item{y}{univariate time series indexed as \code{1,..,T}. Can be either a numeric vector
or matrix, or a class \code{ts} object in which case the dates are correctly passed to
the returned components.}

\item{h}{an integer value giving the horizon of the linear projection. Hamilton recommends
two years, i.e., 8 for quarterly data and 24 for monthly data.}

\item{p}{the number of lags (including the zero lag) used in the linear projection. Should be
chosen so that \code{p > d} where \code{d} is the order of integration of the series. Hamilton
recommends \code{p} to be a multiplicative of one year for seasonal data, i.e., for instance
\code{p=4} for quarterly data and \code{p=12} monthly data.}

\item{x}{object of class 'hfilter' generated by function \code{hfilter}}

\item{...}{currently not in use.}

\item{digits}{number of digits to be printed.}
}
\value{
Returns a class \code{'hfilter'} object containing the following:
\describe{
\item{\code{$cycle}:}{the cyclical component of the series}
\item{\code{$trend}:}{the trend component of the series}
\item{\code{$total}:}{trend + cyclical (shorter than the original seris)}
\item{\code{$beta}:}{the OLS coefficients}
\item{\code{$y}:}{the original series}
\item{\code{$h}:}{the horizon used}
\item{\code{$p}:}{the number of lags (including the zero lag) used}
}
}
\description{
\code{hfilter} separates trend and cycle components of univariate time series
using the method proposed by Hamilton (2018, doi:10.1162/rest_a_00706) based on linear
projections.
}
\details{
The first \code{p + h - 1} values of the series are required to obtain the cycle component
at time \code{p + h} (= first time point of the series). For more details, see the paper
by Hamilton (2018).
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot method

\item \code{print}: print method
}}

\examples{
 data(INDPRO, package="tsfilters")
 IP_filtered <- hfilter(INDPRO, h=24, p=12)
 IP_filtered
 plot(IP_filtered)
}
\references{
\itemize{
\item J.D. Hamilton. 2018. WHY YOU SHOULD NEVER USE THE HODRICK-PRESCOTT FILTER.
\emph{The Review of Economics and Statistics}, 100(5): 831-843.
}
}
