% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrendfilters.R, R/plotmethods.R,
%   R/printmethods.R
\name{hpfilter}
\alias{hpfilter}
\alias{plot.hpfilter}
\alias{print.hpfilter}
\title{Hodrick-Prescott filter}
\usage{
hpfilter(y, lambda = 1600, type = c("one-sided", "two-sided"))

\method{plot}{hpfilter}(x, ...)

\method{print}{hpfilter}(x, ..., digits = 3)
}
\arguments{
\item{y}{univariate time series indexed as \code{1,..,T}. Can be either a numeric vector
or matrix, or a class \code{ts} object in which case the dates are correctly passed to
the returned components.}

\item{lambda}{real number giving the smoothing parameter lambda}

\item{type}{should "one-sided" or "two-sided" HP filter be used? One sided runs the two-sided filter
consecutively for each t (starting from the third observation) and takes the last value as the trend/cycle
component at the time t.}

\item{x}{object of class 'hpfilter' generated by function \code{hpfilter}}

\item{...}{currently not in use.}

\item{digits}{number of digits to be printed.}
}
\value{
Returns a class \code{'hpfilter'} object containing the following:
\describe{
\item{\code{$cycle}:}{the cyclical component of the series}
\item{\code{$trend}:}{the trend component of the series}
\item{\code{$total}:}{trend + cyclical (shorter than the original series)}
\item{\code{$lambda}:}{the smoothing parameter lambda}
\item{\code{$type}:}{one-sided or two-sided}
\item{\code{$y}:}{the original series}
}
If the provided series \code{y} is of class \code{ts}, the dates of the decomposed
series will be set accordingly.
}
\description{
\code{hpfilter} Hodrick-Prescott filter for univariate time series
}
\details{
@import stats

Decompose univariate time series to cyclical component and trend component with Hodrick-Prescott filter.
This function directly uses the function \code{hp_filter} from the package \code{lpirfs}. Notice that
the first two observations are lost in the one-sided filter.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot method

\item \code{print}: print method
}}

\examples{
 # Log of quarterly industrial production index
 data(INDPRO, package="tsfilters")
 IPQ <- log(colMeans(matrix(INDPRO, nrow=3)))
 IPQ <- ts(IPQ, start=start(INDPRO), frequency=4)

 # One-sided
 IPQ_hp <- hpfilter(IPQ, lambda=1600, type="one-sided")
 plot(IPQ_hp)
 IPQ_hp

 # Two-sided
 IPQ_hp2 <- hpfilter(IPQ, lambda=1600, type="two-sided")
 plot(IPQ_hp2)
 IPQ_hp2

}
