% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INDPRO}
\alias{INDPRO}
\title{U.S. industrial production index: INDPRO}
\format{
A univariate class 'ts' time series object containing 1221 observations.
}
\source{
\url{https://fred.stlouisfed.org/series/INDPRO}
}
\usage{
INDPRO
}
\description{
A dataset containing monthly, seasonally adjusted U.S. industrial production
index (2012=100) from 1919 January to 2020 September.
}
\keyword{datasets}
