% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrendfilters.R
\name{logdiff}
\alias{logdiff}
\title{Take logarithm and then first differences of a time series.}
\usage{
logdiff(y)
}
\arguments{
\item{y}{univariate time series indexed as \code{1,..,T}. Can be either a numeric vector
or matrix, or a class \code{ts} object in which case the dates are correctly passed to
the returned components.}
}
\value{
Returns a class \code{'ts'} object containing the log-differenced series.
If the provided series \code{y} is of class \code{ts}, the dates of the detrended
series will be set accordingly.
}
\description{
\code{logdiff} logarithmizes and then takes first differences of a
time series.
}
\details{
The first observation of the series are required as the initial value for the
the first differences. The second observation will thereby be the first observation
of the detrended series.
}
\examples{
 data(INDPRO, package="tsfilters")
 IP_logdiff <- logdiff(INDPRO)
 start(INDPRO)
 start(IP_logdiff)
 plot(IP_logdiff)
}
